requirements.txt
kiteconnect
python-telegram-bot
pandas
numpy
requests
scikit-learn
matplotlib 
yfinance
nsetools 
python-dotenv
pytest 